+++ authpass/linux/flutter/generated_plugin_registrant.cc
@@ -9,7 +9,9 @@
 #include <argon2_ffi/argon2_ffi_plugin.h>
 #include <biometric_storage/biometric_storage_plugin.h>
 #include <file_selector_linux/file_selector_plugin.h>
+#include <hotkey_manager/hotkey_manager_plugin.h>
 #include <url_launcher_linux/url_launcher_plugin.h>
+#include <window_manager/window_manager_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
   g_autoptr(FlPluginRegistrar) argon2_ffi_registrar =
@@ -21,7 +23,13 @@ void fl_register_plugins(FlPluginRegistry* registry) {
   g_autoptr(FlPluginRegistrar) file_selector_linux_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "FileSelectorPlugin");
   file_selector_plugin_register_with_registrar(file_selector_linux_registrar);
+  g_autoptr(FlPluginRegistrar) hotkey_manager_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "HotkeyManagerPlugin");
+  hotkey_manager_plugin_register_with_registrar(hotkey_manager_registrar);
   g_autoptr(FlPluginRegistrar) url_launcher_linux_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "UrlLauncherPlugin");
   url_launcher_plugin_register_with_registrar(url_launcher_linux_registrar);
+  g_autoptr(FlPluginRegistrar) window_manager_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "WindowManagerPlugin");
+  window_manager_plugin_register_with_registrar(window_manager_registrar);
 }