+++ authpass/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.9.4 - unreleased
+
+* Keyboard shortcut for copy TOTP (ctrl+t) #262
+* Desktop: system wide shortcuts (ctrl+alt+f for opening search). #262
+
 # 1.9.3 - 2021-08-27
 
 * Allow removing of files from the list of "recent files". [#260](https://github.com/authpass/authpass/pull/260).